refactor: LineResponse, StationResponse DTO -> Service 계층으로 이동
