+++ src/main/java/wooteco/subway/ui/SectionController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import wooteco.subway.service.SectionService;
-import wooteco.subway.ui.dto.LineResponse;
+import wooteco.subway.service.dto.LineResponse;
 import wooteco.subway.ui.dto.SectionDeleteRequest;
 import wooteco.subway.ui.dto.SectionRequest;
